feat: support es6 module

+++ uninstall.js
@@ -12,7 +12,7 @@ const { bailOut } = require('./utils');
 const {
   COMMIT_MSG_HOOK_FILE,
   PACKAGE_NAME_LABEL,
-  COMMIT_MSG_LABLE,
+  COMMIT_MSG_LABEL,
   PROJECT_ROOT,
 } = require('./constants');
 
@@ -25,7 +25,7 @@ const commitMsgFile = path.resolve(git, 'hooks', COMMIT_MSG_HOOK_FILE);
 // Bail out if we don't have commit-msg file, it might be removed manually.
 if (!exists(commitMsgFile)) {
   console.info(
-    `${PACKAGE_NAME_LABEL}: Not found any ${COMMIT_MSG_LABLE} hook, no need to clean the battle field`,
+    `${PACKAGE_NAME_LABEL}: Not found any ${COMMIT_MSG_LABEL} hook, no need to clean the battle field`,
   );
 
   bailOut();
@@ -33,10 +33,19 @@ if (!exists(commitMsgFile)) {
 
 // If we don't have an old commit-msg file, we should just remove the commit-msg hook.
 if (!exists(`${commitMsgFile}.old`)) {
-  fs.unlinkSync(commitMsgFile);
+  // only remove what we created
+  isFileWeCreated(commitMsgFile) && fs.unlinkSync(commitMsgFile);
 } else {
-  // But if we do have an old one it must restore. *DON'T BE EVIL*.
-  fs.writeFileSync(commitMsgFile, fs.readFileSync(`${commitMsgFile}.old`));
+  // But if we do have an old one it must restored. *DON'T BE EVIL*.
+  fs.copyFileSync(`${commitMsgFile}.old`, commitMsgFile);
+
   fs.chmodSync(commitMsgFile, '755');
   fs.unlinkSync(`${commitMsgFile}.old`);
+}
+
+function isFileWeCreated(fp) {
+  const commitMsgFileContent = fs.readFileSync(fp, 'utf-8');
+  const ID = 'commit-msg-linter';
+
+  return commitMsgFileContent.includes(ID);
 }
